@@ -10,13 +10,13 @@
         }
     ],
     "require": {
-        "php": ">=5.3.0",
+        "php": ">=7.1",
         "apache/log4php": "2.3.*",
-        "lf4php/lf4php": "~4.2"
+        "lf4php/lf4php": "dev-feature/make-php-7-1-compatible"
     },
     "require-dev": {
-        "phpunit/phpunit": "3.7.*",
-        "mockery/mockery": "0.9.4"
+        "phpunit/phpunit": "7.0.2",
+        "mockery/mockery": "1.0"
     },
     "autoload": {
         "psr-0": {"lf4php": "src/"}
