@@ -1,25 +1,5 @@
 <?php
-/*
- * Copyright (c) 2014 Janos Szurovecz
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of
- * this software and associated documentation files (the "Software"), to deal in
- * the Software without restriction, including without limitation the rights to
- * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
- * of the Software, and to permit persons to whom the Software is furnished to do
- * so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
+declare(strict_types=1);
 
 namespace lf4php\impl;
 
@@ -38,7 +18,7 @@ class Log4phpMDCAdapter implements MDCAdapter
      * @param string $key
      * @param string $value
      */
-    public function put($key, $value)
+    public function put(string $key, string $value) : void
     {
         LoggerMDC::put($key, $value);
     }
@@ -47,7 +27,7 @@ class Log4phpMDCAdapter implements MDCAdapter
      * @param string $key
      * @return string
      */
-    public function get($key)
+    public function get(string $key) : ?string
     {
         $result = LoggerMDC::get($key);
         return $result === '' ? null : $result;
@@ -56,12 +36,12 @@ class Log4phpMDCAdapter implements MDCAdapter
     /**
      * @param string $key
      */
-    public function remove($key)
+    public function remove(string $key) : void
     {
         LoggerMDC::remove($key);
     }
 
-    public function clear()
+    public function clear() : void
     {
         LoggerMDC::clear();
     }
@@ -69,7 +49,7 @@ class Log4phpMDCAdapter implements MDCAdapter
     /**
      * @return array|null
      */
-    public function getCopyOfContextMap()
+    public function getCopyOfContextMap() : ?array
     {
         return LoggerMDC::getMap();
     }
@@ -77,7 +57,7 @@ class Log4phpMDCAdapter implements MDCAdapter
     /**
      * @param array $contextMap
      */
-    public function setContextMap(array $contextMap)
+    public function setContextMap(array $contextMap) : void
     {
         foreach ($contextMap as $key => $value) {
             $this->put($key, $value);
