@@ -1,26 +1,5 @@
 <?php
-/*
- * Copyright (c) 2014 Janos Szurovecz
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of
- * this software and associated documentation files (the "Software"), to deal in
- * the Software without restriction, including without limitation the rights to
- * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
- * of the Software, and to permit persons to whom the Software is furnished to do
- * so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
+declare(strict_types=1);
 
 namespace lf4php\impl;
 
@@ -39,7 +18,7 @@ final class StaticLoggerBinder
 
     private $loggerFactory;
 
-    public static function init()
+    public static function init() : void
     {
         self::$SINGLETON = new StaticLoggerBinder();
         self::$SINGLETON->loggerFactory = new Log4phpLoggerFactory();
@@ -48,7 +27,7 @@ final class StaticLoggerBinder
     /**
      * @return Log4phpLoggerFactory
      */
-    public function getLoggerFactory()
+    public function getLoggerFactory() : Log4phpLoggerFactory
     {
         return $this->loggerFactory;
     }
